requirements.txt
requests>=2.25.1
click>=7.1.2
APScheduler>=3.6.3
setuptools>=51.1.0
pytest>=6.2.1
Pillow>=8.0.1